feat: add eventListeners and function to number and operator-buttons
The Value of the buttons will be used to store them into the display-
value. With the Number()-Function it will be safe, that the
calculations will be operated with numbers and not strings.

The display will show the numbers, when you click a operator, then the
display will be set back to "0", but the variable "displayValue" will be
set to a empty string, to start a fresh number for the second number.