fix: remove unwanted behavior, when user starts a new calculation after result
If the user has already a result and begins to calculate a new calculation,
the result from before would have been added.
Now, if there is a num1 stored and the operator is null and the user
clicks on another number, num1 will be reset and a new calculation
can begin.

If the user clicks on a operator again, num1 stays set with the result
before and used for the next follow-up calculation.